requirements.txt
streamlit>=1.28.0
boto3>=1.28.0
langchain>=0.1.0
hypothesis>=6.92.0
pytest>=7.4.0
GitPython>=3.1.40
python-dotenv>=1.0.0
